break out deploying code into new package "apply"
Apply is the word used by terraform to mean deploy. I used this instead of deploy to not have so many "deploy" everywhere.

There should be no functionality change from this CL. It just moves things out into a separate package, making it easier for us to understand and modify in the future. There will be more cleanup to come to clean the tests up.

PiperOrigin-RevId: 250281862