allow default resource to handle name being set outside properties block
In cases like firewalls there is no name field in the properties block (instead the name is on each individual firewall rule, which is a list). For DM we still need a single name to use for the resource name, so allow users to set it outside the properties block and we can reuse the same default resource to manage it.

PiperOrigin-RevId: 257853376