Update path constructor for fhir resources
This cloud function was returning "resource not found" errors when trying to retrieve ../resources/Patient/... using v1beta1. Documentation references .../fhir/Patient/... as the location of the Patient record needed. Updating this path constructor eliminates the error and produces the expected results as observed in the function logs and the FHIR Immunizations web app.